% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{AnimatedLine}
\alias{AnimatedLine}
\title{AnimatedLine}
\usage{
AnimatedLine(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/animated-line/}
}
\details{
\itemize{
\item skipAnimation \code{bool} \cr Default is FALSE If true, animations are skipped.
}
}
