% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{AreaPlot}
\alias{AreaPlot}
\title{AreaPlot}
\usage{
AreaPlot(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/area-plot/}
}
\details{
\itemize{
\item onItemClick \code{func} \cr Default is - Callback fired when a line area item is clicked.Signature:function(event: React.MouseEvent<SVGPathElement, MouseEvent>, lineItemIdentifier: LineItemIdentifier) => voidevent The event source of the callback.lineItemIdentifier The line item identifier.
\item skipAnimation \code{bool} \cr Default is FALSE If true, animations are skipped.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
}
}
