% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{BarLabel}
\alias{BarLabel}
\title{BarLabel}
\usage{
BarLabel(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/bar-label/}
}
\details{
\itemize{
\item height \code{number} \cr Default is - Height of the bar this label belongs to.
\item width \code{number} \cr Default is - Width of the bar this label belongs to.
\item x \code{number} \cr Default is - Position in the x-axis of the bar this label belongs to.
\item xOrigin \code{number} \cr Default is - The x-coordinate of the stack this bar label belongs to.
\item y \code{number} \cr Default is - Position in the y-axis of the bar this label belongs to.
\item yOrigin \code{number} \cr Default is - The y-coordinate of the stack this bar label belongs to.
\item placement \verb{'center', 'outside'} \cr Default is 'center' The placement of the bar label. It controls whether the label is rendered in the center or outside the bar.
}
}
