% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{BarPlot}
\alias{BarPlot}
\title{BarPlot}
\usage{
BarPlot(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/bar-plot/}
}
\details{
\itemize{
\item barLabel \verb{'value', func} \cr Default is - If provided, the function will be used to format the label of the bar. It can be set to 'value' to display the current value.
\item barLabelPlacement \verb{'outside', 'inside'} \cr Default is - The placement of the bar label. It controls whether the label is rendered inside or outside the bar.
\item borderRadius \code{number} \cr Default is - Defines the border radius of the bar element.
\item onItemClick \code{func} \cr Default is - Callback fired when a bar item is clicked.Signature:function(event: React.MouseEvent<SVGElement, MouseEvent>, barItemIdentifier: BarItemIdentifier) => voidevent The event source of the callback.barItemIdentifier The bar item identifier.
\item skipAnimation \code{bool} \cr Default is undefined If true, animations are skipped.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
}
}
