% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartDataProvider}
\alias{ChartDataProvider}
\title{ChartDataProvider}
\usage{
ChartDataProvider(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/chart-data-provider/}
}
\details{
\itemize{
\item colors \verb{Array<string>, func} \cr Default is rainbowSurgePalette Color palette used to colorize multiple series.
\item dataset \verb{Array<object>} \cr Default is - An array of objects that can be used to populate series and axes data using their dataKey property.
\item experimentalFeatures \verb{preferStrictDomainInLineCharts?: bool} \cr Default is - Options to enable features planned for the next major.
\item height \code{number} \cr Default is - The height of the chart in px. If not defined, it takes the height of the parent element.
\item id \code{string} \cr Default is - This prop is used to help implement the accessibility logic. If you don't provide this prop. It falls back to a randomly generated id.
\item localeText \code{object} \cr Default is - Localized text for chart components.
\item margin \verb{number,  bottom?: number, left?: number, right?: number, top?: number } \cr Default is - The margin between the SVG and the drawing area. It's used for leaving some space for extra information such as the x- and y-axis or legend.Accepts a number to be used on all sides or an object with the optional properties: top, bottom, left, and right.
\item series \verb{Array<object>} \cr Default is - The array of series to display. Each type of series has its own specificity. Please refer to the appropriate docs page to learn more about it.
\item skipAnimation \code{bool} \cr Default is - If true, animations are skipped. If unset or false, the animations respects the user's prefers-reduced-motion setting.
\item slotProps \code{object} \cr Default is - The props for the slots.
\item slots \code{object} \cr Default is - Slots to customize charts' components.See Slots API below for more details.
\item width \code{number} \cr Default is - The width of the chart in px. If not defined, it takes the width of the parent element.
}
}
