% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsAxisHighlight}
\alias{ChartsAxisHighlight}
\title{ChartsAxisHighlight}
\usage{
ChartsAxisHighlight(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-axis-highlight/}
}
