% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsAxisTooltipContent}
\alias{ChartsAxisTooltipContent}
\title{ChartsAxisTooltipContent}
\usage{
ChartsAxisTooltipContent(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-axis-tooltip-content/}
}
\details{
\itemize{
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
}
}
