% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsClipPath}
\alias{ChartsClipPath}
\title{ChartsClipPath}
\usage{
ChartsClipPath(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-clip-path/}
}
\details{
\itemize{
\item id \code{string} \cr Default is NA The id of the clip path.
\item offset \verb{bottom?: number, left?: number, right?: number, top?: number} \cr Default is NA Offset, in pixels, of the clip path rectangle from the drawing area.A positive value will move the rectangle outside the drawing area.
}
}
