% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsGrid}
\alias{ChartsGrid}
\title{ChartsGrid}
\usage{
ChartsGrid(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-grid/}
}
\details{
\itemize{
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item horizontal \code{bool} \cr Default is NA Displays horizontal grid.
\item vertical \code{bool} \cr Default is NA Displays vertical grid.
}
}
