% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsLegend}
\alias{ChartsLegend}
\title{ChartsLegend}
\usage{
ChartsLegend(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-legend/}
}
\details{
\itemize{
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item direction \verb{'horizontal', 'vertical'} \cr Default is - The direction of the legend layout. The default depends on the chart.
\item onItemClick \code{func} \cr Default is - Callback fired when a legend item is clicked.Signature:function(event: React.MouseEvent<HTMLButtonElement, MouseEvent>, legendItem: SeriesLegendItemContext, index: number) => voidevent The click event.legendItem The legend item data.index The index of the clicked legend item.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
}
}
