% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsReferenceLine}
\alias{ChartsReferenceLine}
\title{ChartsReferenceLine}
\usage{
ChartsReferenceLine(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-reference-line/}
}
\details{
\itemize{
\item axisId \verb{number, string} \cr Default is The \code{id} of the first defined axis. The id of the axis used for the reference value.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item label \code{string} \cr Default is - The label to display along the reference line.
\item labelAlign \verb{'end', 'middle', 'start'} \cr Default is 'middle' The alignment if the label is in the chart drawing area.
\item labelStyle \code{object} \cr Default is - The style applied to the label.
\item lineStyle \code{object} \cr Default is - The style applied to the line.
\item spacing \verb{number,  x?: number, y?: number } \cr Default is  x: 0, y: 5  on a horizontal line and  x: 5, y: 0  on a vertical line. Additional space around the label in px. Can be a number or an object  x, y  to distinguish space with the reference line and space with axes.
\item x \verb{Date, number, string} \cr Default is - The x value associated with the reference line. If defined the reference line will be vertical.
\item y \verb{Date, number, string} \cr Default is - The y value associated with the reference line. If defined the reference line will be horizontal.
}
}
