% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsText}
\alias{ChartsText}
\title{ChartsText}
\usage{
ChartsText(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-text/}
}
\details{
\itemize{
\item text \code{string} \cr Default is - Text displayed.
\item lineHeight \code{number} \cr Default is - Height of a text line (in em).
\item needsComputation \code{bool} \cr Default is FALSE If true, the line width is computed.
\item style \code{object} \cr Default is - Style applied to text elements.
}
}
