% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsWrapper}
\alias{ChartsWrapper}
\title{ChartsWrapper}
\usage{
ChartsWrapper(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-wrapper/}
}
\details{
\itemize{
\item extendVertically \code{bool} \cr Default is \code{FALSE} if the \code{height} prop is set. And \code{TRUE} otherwise. If true, the chart wrapper set height: 100\%.
\item hideLegend \code{bool} \cr Default is FALSE If true, the legend is not rendered.
\item legendDirection \verb{'horizontal', 'vertical'} \cr Default is 'horizontal' The direction of the legend.
\item legendPosition \verb{horizontal?: 'center', 'end', 'start', vertical?: 'bottom', 'middle', 'top'} \cr Default is  horizontal: 'center', vertical: 'bottom'  The position of the legend.
}
}
