% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ChartsXAxis}
\alias{ChartsXAxis}
\title{ChartsXAxis}
\usage{
ChartsXAxis(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/charts-x-axis/}
}
\details{
\itemize{
\item axisId \verb{number, string} \cr Default is - The id of the axis to render. If undefined, it will be the first defined axis.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item disableLine \code{bool} \cr Default is FALSE If true, the axis line is disabled.
\item disableTicks \code{bool} \cr Default is FALSE If true, the ticks are disabled.
\item label \code{string} \cr Default is - The label of the axis.
\item labelStyle \code{object} \cr Default is - The style applied to the axis label.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
\item tickInterval \verb{'auto', array, func} \cr Default is 'auto' Defines which ticks are displayed. Its value can be: - 'auto' In such case the ticks are computed based on axis scale and other parameters. - a filtering function of the form (value, index) => boolean which is available only if the axis has "point" scale. - an array containing the values where ticks should be displayed.
\item tickLabelInterval \verb{'auto', func} \cr Default is 'auto' Defines which ticks get its label displayed. Its value can be: - 'auto' In such case, labels are displayed if they do not overlap with the previous one. - a filtering function of the form (value, index) => boolean. Warning: the index is tick index, not data ones.
\item tickLabelMinGap \code{number} \cr Default is 4 The minimum gap in pixels between two tick labels. If two tick labels are closer than this minimum gap, one of them will be hidden.
\item tickLabelPlacement \verb{'middle', 'tick'} \cr Default is 'middle' The placement of ticks label. Can be the middle of the band, or the tick position. Only used if scale is 'band'.
\item tickLabelStyle \code{object} \cr Default is - The style applied to ticks text.
\item tickMaxStep \code{number} \cr Default is - Maximal step between two ticks. When using time data, the value is assumed to be in ms. Not supported by categorical axis (band, points).
\item tickMinStep \code{number} \cr Default is - Minimal step between two ticks. When using time data, the value is assumed to be in ms. Not supported by categorical axis (band, points).
\item tickNumber \code{number} \cr Default is - The number of ticks. This number is not guaranteed. Not supported by categorical axis (band, points).
\item tickPlacement \verb{'end', 'extremities', 'middle', 'start'} \cr Default is 'extremities' The placement of ticks in regard to the band interval. Only used if scale is 'band'.
\item tickSize \code{number} \cr Default is 6 The size of the ticks.
}
}
