% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ContinuousColorLegend}
\alias{ContinuousColorLegend}
\title{ContinuousColorLegend}
\usage{
ContinuousColorLegend(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/continuous-color-legend/}
}
\details{
\itemize{
\item axisDirection \verb{'x', 'y', 'z'} \cr Default is 'z' The axis direction containing the color configuration to represent.
\item axisId \verb{number, string} \cr Default is The first axis item. The id of the axis item with the color configuration to represent.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item direction \verb{'horizontal', 'vertical'} \cr Default is 'horizontal' The direction of the legend layout.
\item gradientId \code{string} \cr Default is auto-generated id The id for the gradient to use. If not provided, it will use the generated gradient from the axis configuration. The gradientId will be used as fill="url(#gradientId)".
\item labelPosition \verb{'start', 'end', 'extremes'} \cr Default is 'end' Where to position the labels relative to the gradient.
\item maxLabel \verb{func, string} \cr Default is formattedValue The label to display at the maximum side of the gradient. Can either be a string, or a function. If not defined, the formatted maximal value is display.
\item minLabel \verb{func, string} \cr Default is formattedValue The label to display at the minimum side of the gradient. Can either be a string, or a function.
\item reverse \code{bool} \cr Default is FALSE If true, the gradient and labels will be reversed.
\item rotateGradient \code{bool} \cr Default is - If provided, the gradient will be rotated by 90deg. Useful for linear gradients that are not in the correct orientation.
\item thickness \code{number} \cr Default is 12 The thickness of the gradient
}
}
