% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Gauge}
\alias{Gauge}
\title{Gauge}
\usage{
Gauge(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/gauge/}
}
\details{
\itemize{
\item cornerRadius \verb{number, string} \cr Default is 0 The radius applied to arc corners (similar to border radius). Set it to '50\%' to get rounded arc.
\item cx \verb{number, string} \cr Default is - The x coordinate of the arc center. Can be a number (in px) or a string with a percentage such as '50\%'. The '100\%' is the width the drawing area.
\item cy \verb{number, string} \cr Default is - The y coordinate of the arc center. Can be a number (in px) or a string with a percentage such as '50\%'. The '100\%' is the height the drawing area.
\item endAngle \code{number} \cr Default is 360 The end angle (deg).
\item height \code{number} \cr Default is - The height of the chart in px. If not defined, it takes the height of the parent element.
\item id \code{string} \cr Default is - This prop is used to help implement the accessibility logic. If you don't provide this prop. It falls back to a randomly generated id.
\item innerRadius \verb{number, string} \cr Default is '80\%' The radius between circle center and the beginning of the arc. Can be a number (in px) or a string with a percentage such as '50\%'. The '100\%' is the maximal radius that fit into the drawing area.
\item margin \verb{number,  bottom?: number, left?: number, right?: number, top?: number } \cr Default is - The margin between the SVG and the drawing area. It's used for leaving some space for extra information such as the x- and y-axis or legend.Accepts a number to be used on all sides or an object with the optional properties: top, bottom, left, and right.
\item outerRadius \verb{number, string} \cr Default is '100\%' The radius between circle center and the end of the arc. Can be a number (in px) or a string with a percentage such as '50\%'. The '100\%' is the maximal radius that fit into the drawing area.
\item skipAnimation \code{bool} \cr Default is - If true, animations are skipped. If unset or false, the animations respects the user's prefers-reduced-motion setting.
\item startAngle \code{number} \cr Default is 0 The start angle (deg).
\item value \code{number} \cr Default is - The value of the gauge. Set to null to not display a value.
\item valueMax \code{number} \cr Default is 100 The maximal value of the gauge.
\item valueMin \code{number} \cr Default is 0 The minimal value of the gauge.
\item width \code{number} \cr Default is - The width of the chart in px. If not defined, it takes the width of the parent element.
}
}
