% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{GaugeValueText}
\alias{GaugeValueText}
\title{GaugeValueText}
\usage{
GaugeValueText(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/gauge-value-text/}
}
\details{
\itemize{
\item text \verb{func, string} \cr Default is NA The text to display in the gauge. Can be a string or a function receiving the value and returning a string.
}
}
