% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{MarkElement}
\alias{MarkElement}
\title{MarkElement}
\usage{
MarkElement(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/mark-element/}
}
\details{
\itemize{
\item dataIndex \code{number} \cr Default is - The index to the element in the series' data array.
\item shape \verb{'circle', 'cross', 'diamond', 'square', 'star', 'triangle', 'wye'} \cr Default is - The shape of the marker.
\item isFaded \code{bool} \cr Default is FALSE If true, the marker is faded.
\item isHighlighted \code{bool} \cr Default is FALSE If true, the marker is highlighted.
\item skipAnimation \code{bool} \cr Default is - If true, animations are skipped.
}
}
