% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muiChartsDependency.R
\name{muiChartsDependency}
\alias{muiChartsDependency}
\title{Mui Charts UI JS dependency}
\usage{
muiChartsDependency()
}
\value{
HTML dependency object.
}
\description{
Mui Charts UI JS dependency
}
