% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{PieArc}
\alias{PieArc}
\title{PieArc}
\usage{
PieArc(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/pie-arc/}
}
\details{
\itemize{
\item skipAnimation \code{bool} \cr Default is NA If true, the animation is disabled.
\item skipInteraction \code{bool} \cr Default is NA If true, the default event handlers are disabled. Those are used, for example, to display a tooltip or highlight the arc on hover.
}
}
