% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{PieArcLabelPlot}
\alias{PieArcLabelPlot}
\title{PieArcLabelPlot}
\usage{
PieArcLabelPlot(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/pie-arc-label-plot/}
}
\details{
\itemize{
\item id \verb{number, string} \cr Default is - The id of this series.
\item outerRadius \code{number} \cr Default is - The radius between circle center and the end of the arc.
\item arcLabel \verb{'formattedValue', 'label', 'value', func} \cr Default is - The label displayed into the arc.
\item arcLabelMinAngle \code{number} \cr Default is 0 The minimal angle required to display the arc label.
\item arcLabelRadius \code{number} \cr Default is (innerRadius - outerRadius) / 2 The radius between circle center and the arc label in px.
\item cornerRadius \code{number} \cr Default is 0 The radius applied to arc corners (similar to border radius).
\item faded \verb{additionalRadius?: number, arcLabelRadius?: number, color?: string, cornerRadius?: number, innerRadius?: number, outerRadius?: number, paddingAngle?: number} \cr Default is  additionalRadius: -5  Override the arc attributes when it is faded.
\item highlighted \verb{additionalRadius?: number, arcLabelRadius?: number, color?: string, cornerRadius?: number, innerRadius?: number, outerRadius?: number, paddingAngle?: number} \cr Default is - Override the arc attributes when it is highlighted.
\item innerRadius \code{number} \cr Default is 0 The radius between circle center and the beginning of the arc.
\item paddingAngle \code{number} \cr Default is 0 The padding angle (deg) between two arcs.
\item skipAnimation \code{bool} \cr Default is FALSE If true, animations are skipped.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
}
}
