% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{PiecewiseColorLegend}
\alias{PiecewiseColorLegend}
\title{PiecewiseColorLegend}
\usage{
PiecewiseColorLegend(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/piecewise-color-legend/}
}
\details{
\itemize{
\item axisDirection \verb{'x', 'y', 'z'} \cr Default is 'z' The axis direction containing the color configuration to represent.
\item axisId \verb{number, string} \cr Default is The first axis item. The id of the axis item with the color configuration to represent.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item direction \verb{'horizontal', 'vertical'} \cr Default is 'horizontal' The direction of the legend layout.
\item labelFormatter \code{func} \cr Default is - Format the legend labels.Signature:function(params: PiecewiseLabelFormatterParams) => string | nullparams The bound of the piece to format.
\item labelPosition \verb{'start', 'end', 'extremes', 'inline-start', 'inline-end'} \cr Default is 'extremes' Where to position the labels relative to the gradient.
\item markType \verb{'square', 'circle', 'line'} \cr Default is 'square' The type of the mark.
\item onItemClick \code{func} \cr Default is - Callback fired when a legend item is clicked.Signature:function(event: React.MouseEvent<HTMLButtonElement, MouseEvent>, legendItem: PiecewiseColorLegendItemContext, index: number) => voidevent The click event.legendItem The legend item data.index The index of the clicked legend item.
}
}
