% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{PieChart}
\alias{PieChart}
\title{PieChart}
\usage{
PieChart(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/pie-chart/}
}
\details{
\itemize{
\item series \verb{Array<object>} \cr Default is - The series to display in the pie chart. An array of PieSeries objects.
\item colors \verb{Array<string>, func} \cr Default is rainbowSurgePalette Color palette used to colorize multiple series.
\item dataset \verb{Array<object>} \cr Default is - An array of objects that can be used to populate series and axes data using their dataKey property.
\item height \code{number} \cr Default is - The height of the chart in px. If not defined, it takes the height of the parent element.
\item hideLegend \code{bool} \cr Default is - If true, the legend is not rendered.
\item highlightedItem \verb{dataIndex?: number, seriesId: number, string} \cr Default is - The highlighted item. Used when the highlight is controlled.
\item id \code{string} \cr Default is - This prop is used to help implement the accessibility logic. If you don't provide this prop. It falls back to a randomly generated id.
\item loading \code{bool} \cr Default is FALSE If true, a loading overlay is displayed.
\item localeText \code{object} \cr Default is - Localized text for chart components.
\item margin \verb{number,  bottom?: number, left?: number, right?: number, top?: number } \cr Default is - The margin between the SVG and the drawing area. It's used for leaving some space for extra information such as the x- and y-axis or legend.Accepts a number to be used on all sides or an object with the optional properties: top, bottom, left, and right.
\item onHighlightChange \code{func} \cr Default is - The callback fired when the highlighted item changes.Signature:function(highlightedItem: HighlightItemData | null) => voidhighlightedItem The newly highlighted item.
\item onItemClick \code{func} \cr Default is - Callback fired when a pie arc is clicked.
\item showToolbar \code{bool} \cr Default is FALSE If true, shows the default chart toolbar.
\item skipAnimation \code{bool} \cr Default is - If true, animations are skipped. If unset or false, the animations respects the user's prefers-reduced-motion setting.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
\item width \code{number} \cr Default is - The width of the chart in px. If not defined, it takes the width of the parent element.
}
}
