% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{RadarAxis}
\alias{RadarAxis}
\title{RadarAxis}
\usage{
RadarAxis(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/radar-axis/}
}
\details{
\itemize{
\item angle \code{number} \cr Default is - The absolute rotation angle of the metrics (in degree) If not defined the metric angle will be used.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item divisions \code{number} \cr Default is 1 The number of divisions with label.
\item dominantBaseline \verb{'alphabetic', 'auto', 'central', 'hanging', 'ideographic', 'inherit', 'mathematical', 'middle', 'no-change', 'reset-size', 'text-after-edge', 'text-before-edge', 'use-script', func} \cr Default is - The labels dominant baseline or a function returning the dominant baseline for a given axis angle (in degree).
\item labelOrientation \verb{'horizontal', 'rotated'} \cr Default is 'horizontal' Defines how label align with the axis. - 'horizontal': labels stay horizontal and their placement change with the axis angle. - 'rotated': labels are rotated 90deg relatively to their axis.
\item metric \code{string} \cr Default is - The metric to get. If undefined, the hook returns null
\item textAnchor \verb{'end', 'inherit', 'middle', 'start', func} \cr Default is - The labels text anchor or a function returning the text anchor for a given axis angle (in degree).
}
}
