% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{RadarChart}
\alias{RadarChart}
\title{RadarChart}
\usage{
RadarChart(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/radar-chart/}
}
\details{
\itemize{
\item radar \verb{labelFormatter?: func, labelGap?: number, max?: number, metrics: Array<string>, Array< max?: number, min?: number, name: string >, startAngle?: number} \cr Default is - The configuration of the radar scales.
\item series \verb{Array<object>} \cr Default is - The series to display in the bar chart. An array of RadarSeries objects.
\item colors \verb{Array<string>, func} \cr Default is rainbowSurgePalette Color palette used to colorize multiple series.
\item disableAxisListener \code{bool} \cr Default is FALSE If true, the charts will not listen to the mouse move event. It might break interactive features, but will improve performance.
\item divisions \code{number} \cr Default is 5 The number of divisions in the radar grid.
\item height \code{number} \cr Default is - The height of the chart in px. If not defined, it takes the height of the parent element.
\item hideLegend \code{bool} \cr Default is - If true, the legend is not rendered.
\item highlight \verb{'axis', 'none', 'series'} \cr Default is 'axis' Indicates if the chart should highlight items per axis or per series.
\item highlightedItem \verb{dataIndex?: number, seriesId: number, string} \cr Default is - The highlighted item. Used when the highlight is controlled.
\item id \code{string} \cr Default is - This prop is used to help implement the accessibility logic. If you don't provide this prop. It falls back to a randomly generated id.
\item loading \code{bool} \cr Default is FALSE If true, a loading overlay is displayed.
\item localeText \code{object} \cr Default is - Localized text for chart components.
\item margin \verb{number,  bottom?: number, left?: number, right?: number, top?: number } \cr Default is - The margin between the SVG and the drawing area. It's used for leaving some space for extra information such as the x- and y-axis or legend.Accepts a number to be used on all sides or an object with the optional properties: top, bottom, left, and right.
\item onAreaClick \code{func} \cr Default is - Callback fired when an area is clicked.Signature:function(event: React.MouseEvent<SVGPathElement, MouseEvent>, radarItemIdentifier: RadarItemIdentifier) => voidevent The event source of the callback.radarItemIdentifier The radar item identifier.
\item onAxisClick \code{func} \cr Default is - The function called for onClick events. The second argument contains information about all line/bar elements at the current mouse position.Signature:function(event: MouseEvent, data: null | ChartsAxisData) => voidevent The mouse event recorded on the \if{html}{\out{<svg/>}} element.data The data about the clicked axis and items associated with it.
\item onHighlightChange \code{func} \cr Default is - The callback fired when the highlighted item changes.Signature:function(highlightedItem: HighlightItemData | null) => voidhighlightedItem The newly highlighted item.
\item onMarkClick \code{func} \cr Default is - Callback fired when a mark is clicked.Signature:function(event: React.MouseEvent<SVGPathElement, MouseEvent>, radarItemIdentifier: RadarItemIdentifier) => voidevent The event source of the callback.radarItemIdentifier The radar item identifier.
\item shape \verb{'circular', 'sharp'} \cr Default is 'sharp' The grid shape.
\item showToolbar \code{bool} \cr Default is FALSE If true, shows the default chart toolbar.
\item skipAnimation \code{bool} \cr Default is - If true, animations are skipped. If unset or false, the animations respects the user's prefers-reduced-motion setting.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
\item stripeColor \code{func} \cr Default is (index) => index \% 2 === 1 ? (theme.vars || theme).palette.text.secondary : 'none' Get stripe fill color. Set it to null to remove stripesSignature:function(index: number) => stringindex The index of the stripe band.
\item width \code{number} \cr Default is - The width of the chart in px. If not defined, it takes the width of the parent element.
}
}
