% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{RadarGrid}
\alias{RadarGrid}
\title{RadarGrid}
\usage{
RadarGrid(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/radar-grid/}
}
\details{
\itemize{
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item divisions \code{number} \cr Default is 5 The number of divisions in the radar grid.
\item shape \verb{'circular', 'sharp'} \cr Default is 'sharp' The grid shape.
\item stripeColor \code{func} \cr Default is (index) => index \% 2 === 1 ? (theme.vars || theme).palette.text.secondary : 'none' Get stripe fill color. Set it to null to remove stripesSignature:function(index: number) => stringindex The index of the stripe band.
}
}
