% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{RadarSeriesMarks}
\alias{RadarSeriesMarks}
\title{RadarSeriesMarks}
\usage{
RadarSeriesMarks(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/charts/radar-series-marks/}
}
\details{
\itemize{
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item onItemClick \code{func} \cr Default is NA Callback fired when a mark is clicked.Signature:function(event: React.MouseEvent<SVGPathElement, MouseEvent>, radarItemIdentifier: RadarItemIdentifier) => voidevent The event source of the callback.radarItemIdentifier The radar item identifier.
\item seriesId \code{string} \cr Default is NA The id of the series to display. If undefined all series are displayed.
}
}
