% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_films}
\alias{starwars_films}
\title{Star Wars Films}
\format{
A data frame with 6 rows and 10 variables:
\describe{
  \item{title}{\code{character}. The title of the film.}
  \item{episode_id}{\code{integer}. The episode number of the film.}
  \item{director}{\code{character}. The name of the director of the film.}
  \item{producer}{\code{character}. The name(s) of the producer(s) of the film.}
  \item{release_date}{\code{Date}. The release date of the film in ISO 8601 format.}
  \item{characters}{\code{list} of \code{character} vectors. A list of character names appearing in the film.}
  \item{planets}{\code{list} of \code{character} vectors. A list of planet names appearing in the film.}
  \item{starships}{\code{list} of \code{character} vectors. A list of starship names appearing in the film.}
  \item{vehicles}{\code{list} of \code{character} vectors. A list of vehicle names appearing in the film.}
  \item{species}{\code{list} of \code{character} vectors. A list of species names appearing in the film.}
}
}
\source{
\url{https://swapi.info/api/films}
}
\usage{
starwars_films
}
\description{
A dataset containing information about the Star Wars films, fetched from the
Star Wars API (SWAPI).
}
\keyword{datasets}
