% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_people}
\alias{starwars_people}
\title{Star Wars People}
\format{
A data frame with 82 rows and 11 variables:
\describe{
  \item{name}{\code{character}. The name of the character.}
  \item{height}{\code{integer}. The height of the character in centimeters.}
  \item{mass}{\code{double}. The mass of the character in kilograms.}
  \item{hair_color}{\code{character}. The hair color of the character.}
  \item{skin_color}{\code{character}. The skin color of the character.}
  \item{eye_color}{\code{character}. The eye color of the character.}
  \item{birth_year}{\code{double}. The birth year of the character relative to the Battle of
    Yavin (BBY). A higher number means earlier in the timeline.}
  \item{gender}{\code{character}. The gender of the character.}
  \item{films}{\code{list} of \code{character} vectors. A list of film titles this character appeared in.}
  \item{starships}{\code{list} of \code{character} vectors. A list of starship names this character has piloted.}
  \item{vehicles}{\code{list} of \code{character} vectors. A list of vehicle names this character has piloted.}
}
}
\source{
\url{https://swapi.info/api/people}
}
\usage{
starwars_people
}
\description{
A dataset containing information about people (characters) in the Star Wars
universe, fetched from the Star Wars API (SWAPI).
}
\keyword{datasets}
