% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_planets}
\alias{starwars_planets}
\title{Star Wars Planets}
\format{
A data frame with 60 rows and 11 variables:
\describe{
  \item{name}{\code{character}. The name of the planet.}
  \item{rotation_period}{\code{integer}. The number of standard hours it takes for the planet
    to complete a single rotation on its axis.}
  \item{orbital_period}{\code{integer}. The number of standard days it takes for the planet
    to orbit once around the local star.}
  \item{diameter}{\code{integer}. The diameter of the planet in kilometers.}
  \item{climate}{\code{character}. The climate(s) of the planet.}
  \item{gravity}{\code{character}. The gravity of the planet relative to standard gravity.}
  \item{terrain}{\code{character}. The terrain type(s) found on the planet.}
  \item{surface_water}{\code{double}. The percentage of the planet surface covered by water
    or other bodies of water.}
  \item{population}{\code{double}. The average population of sentient beings inhabiting
    the planet.}
  \item{residents}{\code{list} of \code{character} vectors. A list of character names who are residents of the planet.}
  \item{films}{\code{list} of \code{character} vectors. A list of film titles in which this planet appears.}
}
}
\source{
\url{https://swapi.info/api/planets}
}
\usage{
starwars_planets
}
\description{
A dataset containing information about planets in the Star Wars universe,
fetched from the Star Wars API (SWAPI).
}
\keyword{datasets}
