% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_species}
\alias{starwars_species}
\title{Star Wars Species}
\format{
A data frame with 37 rows and 12 variables:
\describe{
  \item{name}{\code{character}. The name of the species.}
  \item{classification}{\code{character}. The classification of the species (e.g. mammal, reptile).}
  \item{designation}{\code{character}. The designation of the species (e.g. sentient, reptilian).}
  \item{average_height}{\code{double}. The average height of the species in centimeters.}
  \item{average_lifespan}{\code{double}. The average lifespan of the species in years.}
  \item{skin_colors}{\code{character}. A comma-separated string of common skin colors for
    the species.}
  \item{hair_colors}{\code{character}. A comma-separated string of common hair colors for
    the species.}
  \item{eye_colors}{\code{character}. A comma-separated string of common eye colors for
    the species.}
  \item{language}{\code{character}. The language commonly spoken by the species.}
  \item{homeworld}{\code{character}. The name of the planet this species originates from.}
  \item{people}{\code{list} of \code{character} vectors. A list of character names that are members of this species.}
  \item{films}{\code{list} of \code{character} vectors. A list of film titles in which this species appears.}
}
}
\source{
\url{https://swapi.info/api/species}
}
\usage{
starwars_species
}
\description{
A dataset containing information about species in the Star Wars universe,
fetched from the Star Wars API (SWAPI).
}
\keyword{datasets}
