% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_starships}
\alias{starwars_starships}
\title{Star Wars Starships}
\format{
A data frame with 36 rows and 15 variables:
\describe{
  \item{name}{\code{character}. The name of the starship.}
  \item{model}{\code{character}. The model or official name of the starship.}
  \item{manufacturer}{\code{character}. The manufacturer of the starship.}
  \item{cost_in_credits}{\code{double}. The cost of the starship new, in galactic credits.}
  \item{length}{\code{double}. The length of the starship in meters.}
  \item{max_atmosphering_speed}{\code{double}. The maximum speed of the starship in atmosphere.
    \code{NA} if the starship is incapable of atmospheric flight.}
  \item{crew}{\code{double}. The number of personnel needed to run or pilot the starship.}
  \item{passengers}{\code{double}. The number of non-essential people this starship can transport.}
  \item{cargo_capacity}{\code{double}. The maximum cargo load in kilograms.}
  \item{consumables}{\code{character}. The maximum length of time the starship can provide
    consumables for its entire crew without resupplying.}
  \item{hyperdrive_rating}{\code{double}. The class of the starship's hyperdrive.}
  \item{MGLT}{\code{double}. The maximum number of Megalights this starship can travel
    in a standard hour.}
  \item{starship_class}{\code{character}. The class of the starship.}
  \item{pilots}{\code{list} of \code{character} vectors. A list of character names who have piloted this starship.}
  \item{films}{\code{list} of \code{character} vectors. A list of film titles in which this starship appears.}
}
}
\source{
\url{https://swapi.info/api/starships}
}
\usage{
starwars_starships
}
\description{
A dataset containing information about starships in the Star Wars universe,
fetched from the Star Wars API (SWAPI).
}
\keyword{datasets}
